requirements.txt
django
django-cors-headers
django-countries
django-environ
django-filter
django-timezone-field
django-tinymce
djangorestframework
djangorestframework-simplejwt
drf-spectacular[sidecar]
drf-spectacular-sidecar
openpyxl==2.4.0
phonenumbers
pillow
psycopg2-binary
celery
requests
markdown
setuptools